fix: :recycle: Change init structure
