+++ apps/server/src/init/db/index.js
@@ -1,0 +1,16 @@
+import setupProjects from '../db/setup-projects.js'
+import setupUsers from '../db/setup-users.js'
+import { createOrganization } from '../../models/queries/organization-queries.js'
+import { dropTables, synchroniseModels } from '../../connect.js'
+
+export const initDb = async (data) => {
+  await dropTables()
+  await synchroniseModels()
+
+  for (const org of data.organizations) {
+    await createOrganization(org)
+  }
+
+  await setupUsers(data.users)
+  await setupProjects(data.projects)
+}
