+++ apps/server/src/init/db/setup-users.js
@@ -1,5 +1,5 @@
-import app from '../app.js'
-import { createUser } from '../models/queries/user-queries.js'
+import app from '../../app.js'
+import { createUser } from '../../models/queries/user-queries.js'
 
 export default async (users) => {
   app.log.info('Creating users...')
