+++ apps/server/src/init/services/index.js
@@ -1,0 +1,9 @@
+import { init, project, repository, environment } from '../../connectors/index.js'
+export const initServices = async () => {
+  await init()
+  await project.create('dinum', 'temp', 'bisounours@gouv.fr')
+  await project.create('dinum', 'toto', 'bisounours@gouv.fr')
+  await project.delete('dinum', 'temp', 'bisounours@gouv.fr')
+  await environment.create({ organization: 'dinum', project: 'toto', infraRepos: [], environment: 'prod' })
+  await repository.create({ name: 'rope', organization: 'dinum', project: 'toto' })
+}
